move login\register screens to /auth folder & now use 'AppLoadingIndicator' on 'SplashScreen' Refs #17
