+++ frontend/lib/presentation/screens/auth/register_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import '../../logic/services/providers.dart';
+import '../../../logic/services/providers.dart';
 
 
 class RegisterScreen extends ConsumerWidget {
